pyproject.toml
[tool.poetry]
name = "anna"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = "^0.16.14"
fastapi = "*"
uvicorn = "^0.11.8"
python-dotenv = "^0.14.0"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"


[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "*"
isort = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-xdist = "*"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

